Clarify  motion path and position interaction
If offset-position is a position, the motion path translation (and
hence left/right/top/bottom if position is 'absolute') can be observed
if 'rotate' or 'scale' are set.
https://drafts.csswg.org/css-transforms-2/#ctm

If offset-path is a geometry-box, or a basic shape (other than a circle
or ellipse with implicit center), then the value of offset-position is
ignored and we have absolute positioning even if offset-position is
'auto' and position is 'static'.
https://drafts.fxtf.org/motion-1/#offset-path-property

offset-position and offset-anchor do not support the deprecated 3-value
syntax used by 'background-position'
https://lists.w3.org/Archives/Public/www-style/2017Mar/0054.html

position 'static' (the initial value) does not cause offset-position to
be ignored, just as it does not cause offset-path to be ignored.

resolves #45
resolves #77