[motion-1] Clarify meaning of contain
'smallest clipped area' gave no guidance when the path and the box
had zero area in common regarless of the value of offset-distance.

For example, this can occur when the path size is 0, or when
offset-anchor is substantially outside the box.

We now define the meaning of 'contain' by specifying that the path
size is minimally increased so that the box can be contained.

resolves #22