Fix the filtering description to get opacity right
With the previous description, opacity was implicitly included in the final set of filters applied to all content. But that's incorrect, at least according to expectation and behavior. Opacity is special here - it needs to be separately applied to the filtered backdrop, because if it is applied only at the end, the fully-opaque contents of B will have already painted over the filtered backdrop. There won't be any show-through of the backdrop, even though the expectation is to see it, given that opacity is less than 1.
While that change is being made, I've also copied the other effects and clips to the backdrop-filtered image as well. These could possibly be defined to only apply once at the end, but the resulting image will be identical or nearly-so, and in practice, the implementation is typically closer to this new definition.